Alerting: Refactor integration tests (#99519)
---------

(cherry picked from commit af663dadc7c551d53e3de36b2ecf89aa81174a41)
